Frontend/folder restructure (#65)
* complete folder restructure of the project

* rest of the folder setup

* complete folder restructure of the project

* rest of the folder setup

* fix for isLoggedIn default state
