--- frontend/src/Notifications.jsx
@@ -1,14 +1,0 @@
-import { API_URL } from './config';
-import React from 'react';
-import './App.css';
-
-function Notifications() {
-    return (
-        <div className="App">
-            <div className="Center-content-header"><h1>Notifications</h1></div>
-            <div className="Center-content">Coming soon...</div>
-        </div>
-    );
-}
-
-export default Notifications;
