--- frontend/src/Polls.jsx
@@ -1,16 +1,0 @@
-import { API_URL } from './config';
-import React from 'react';
-import './App.css';
-
-function Polls() {
-
-    return (
-        <div className="App">
-            <div className="Center-content-header"><h1>Polls</h1></div>
-            <div className="Center-content">Coming Soon...</div>
-        </div>
-    );
-
-}
-
-export default Polls;
