+++ frontend/src/components/marketDetails/MarketDetails.css
