+++ frontend/src/components/polls/Polls.jsx
@@ -1,0 +1,16 @@
+import { API_URL } from '../../config';
+import React from 'react';
+import '../../App.css';
+
+function Polls() {
+  return (
+    <div className='App'>
+      <div className='Center-content-header'>
+        <h1>Polls</h1>
+      </div>
+      <div className='Center-content'>Coming Soon...</div>
+    </div>
+  );
+}
+
+export default Polls;
