+++ frontend/src/components/profileEdit/ProfileEdit.jsx
@@ -1,0 +1,22 @@
+// EmojiSelector.js
+import React, { useState } from 'react';
+
+const ProfileEdit = ({ onSelectEmoji }) => {
+  const emojis = ['😀', '😃', '😄', '😁', '😆', '😅', '😂', '🤣', '😊', '😇'];
+
+  const handleEmojiSelect = (emoji) => {
+    onSelectEmoji(emoji);
+  };
+
+  return (
+    <div>
+      {emojis.map((emoji) => (
+        <button key={emoji} onClick={() => handleEmojiSelect(emoji)}>
+          {emoji}
+        </button>
+      ))}
+    </div>
+  );
+};
+
+export default ProfileEdit;
